Adjust ci pipelines
Now there's a pipeline for PRs, and the one for merges to the main
branch will not only run the tests, but also deploy in prod if they
pass, and then run them in prod if deployment succeeds.

Automatic deployment to vercel will not happen anymore on pushes to
the main branch. This will only happen if e2e tests pass.